Removed VariantAnnotationWriter. Added new constructor to parent class. Moved parent class to writer package.
